added coupon pages

+++ database/migrations/2018_08_15_091130_create_home_page_images_table.php
@@ -20,7 +20,7 @@ class CreateHomePageImagesTable extends Migration
             $table->tinyInteger('position')->unsigned();
             $table->smallInteger('width')->unsigned();
             $table->smallInteger('height')->unsigned();
-            $table->tinyInteger('sort_order')->unsigned();
+            $table->smallInteger('sort_order')->unsigned();
             $table->timestamps();
         });
     }
